refactor: split cycle history modal modules for CI hygiene
