+++ src/components/CycleManager/CycleHistoryModal.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useRef, useState } from 'react';
+import React, { useRef, useState } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import {
   selectSavedCycles,
@@ -13,6 +13,7 @@ import './CycleHistoryModal.css';
 import ConfirmationModal from '../DrawingTools/ConfirmationModal';
 import CycleHistoryModalDialog from './CycleHistoryModalDialog';
 import { deferCloseAfterConfirm } from './cycleHistoryModalUtils';
+import { useCycleHistoryModalKeyboard } from './useCycleHistoryModalKeyboard';
 
 interface CycleHistoryModalProps {
   isOpen: boolean;
@@ -37,75 +38,17 @@ const CycleHistoryModal: React.FC<CycleHistoryModalProps> = ({ isOpen, onClose }
     onConfirm: () => void;
   } | null>(null);
 
-  const getFocusableElements = (root: HTMLElement | null): HTMLElement[] => {
-    if (!root) return [];
-    return Array.from(
-      root.querySelectorAll<HTMLElement>(
-        'button, [href], input, select, textarea, [tabindex]:not([tabindex="-1"])',
-      ),
-    );
-  };
-
-  const handleTabNavigation = React.useCallback((event: KeyboardEvent, root: HTMLElement | null) => {
-    if (!root) return;
-    const focusable = getFocusableElements(root);
-    if (focusable.length === 0) return;
-
-    const first = focusable[0];
-    const last = focusable[focusable.length - 1];
-    const isFirstActive = document.activeElement === first;
-    const isLastActive = document.activeElement === last;
-
-    if (event.shiftKey && isFirstActive) {
-      event.preventDefault();
-      last.focus();
-    } else if (!event.shiftKey && isLastActive) {
-      event.preventDefault();
-      first.focus();
-    }
-  }, []);
-
-  const handleModalKeyDown = React.useCallback(
-    (event: KeyboardEvent) => {
-      const isEscape = event.key === 'Escape';
-      const isTab = event.key === 'Tab';
-      const canTab = Boolean(modalRef.current && !confirmAction);
-
-      if (isEscape) {
-        if (confirmAction) {
-          setConfirmAction(null);
-        } else {
-          onClose();
-        }
-        return;
-      }
-
-      if (isTab && canTab) {
-        handleTabNavigation(event, modalRef.current);
-      }
-    },
-    [onClose, confirmAction, handleTabNavigation],
-  );
-
-  useEffect(() => {
-    if (!isOpen) return;
-
-    const syncFocus = () => {
-      if (!modalRef.current) return;
-      const focusable = getFocusableElements(modalRef.current);
-      focusable[0]?.focus();
-    };
-
-    window.addEventListener('keydown', handleModalKeyDown);
-    syncFocus();
-
-    return () => {
-      window.removeEventListener('keydown', handleModalKeyDown);
-    };
-  }, [isOpen, handleModalKeyDown]);
+  useCycleHistoryModalKeyboard({
+    isOpen,
+    modalRef,
+    confirmAction,
+    onClose,
+    onDismissConfirm: () => setConfirmAction(null),
+  });
 
   if (!isOpen) return null;
 
+  /** Persists the current forecast cycle with an optional label. */
   const handleSaveCurrent = () => {
     dispatch(saveCurrentCycle({ label: newLabel.trim() || undefined }));
     setNewLabel('');
@@ -113,6 +56,7 @@ const CycleHistoryModal: React.FC<CycleHistoryModalProps> = ({ isOpen, onClose }
     addToast('Cycle saved successfully!', 'success');
   };
 
+  /** Prompts before replacing the active cycle with a saved one. */
   const handleLoadCycle = (cycleId: string) => {
     setConfirmAction({
       title: 'Load Cycle',
@@ -126,6 +70,7 @@ const CycleHistoryModal: React.FC<CycleHistoryModalProps> = ({ isOpen, onClose }
     });
   };
 
+  /** Prompts before deleting a saved cycle permanently. */
   const handleDeleteCycle = (cycleId: string) => {
     setConfirmAction({
       title: 'Delete Cycle',
@@ -157,6 +102,27 @@ const CycleHistoryModal: React.FC<CycleHistoryModalProps> = ({ isOpen, onClose }
     handleDeleteCycle(cycleId);
   };
 
+  /** Reveals the optional-label save form. */
+  const handleOpenSaveForm = () => {
+    setShowSaveForm(true);
+  };
+
+  /** Updates the optional label for the cycle being saved. */
+  const handleNewLabelChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+    setNewLabel(event.target.value);
+  };
+
+  /** Hides the save form and clears the draft label. */
+  const handleCancelSaveForm = () => {
+    setShowSaveForm(false);
+    setNewLabel('');
+  };
+
+  /** Dismisses the nested confirmation dialog without acting. */
+  const handleCancelConfirm = () => {
+    setConfirmAction(null);
+  };
+
   return (
     <ModalPortal>
       <CycleHistoryModalDialog
@@ -166,13 +132,10 @@ const CycleHistoryModal: React.FC<CycleHistoryModalProps> = ({ isOpen, onClose }
         showSaveForm={showSaveForm}
         newLabel={newLabel}
         onClose={onClose}
-        onOpenSaveForm={() => setShowSaveForm(true)}
-        onNewLabelChange={(event) => setNewLabel(event.target.value)}
+        onOpenSaveForm={handleOpenSaveForm}
+        onNewLabelChange={handleNewLabelChange}
         onSaveCurrent={handleSaveCurrent}
-        onCancelSaveForm={() => {
-          setShowSaveForm(false);
-          setNewLabel('');
-        }}
+        onCancelSaveForm={handleCancelSaveForm}
         onLoadClick={handleCycleLoadClick}
         onDeleteClick={handleCycleDeleteClick}
       />
@@ -182,7 +145,7 @@ const CycleHistoryModal: React.FC<CycleHistoryModalProps> = ({ isOpen, onClose }
           title={confirmAction.title}
           message={confirmAction.message}
           onConfirm={confirmAction.onConfirm}
-          onCancel={() => setConfirmAction(null)}
+          onCancel={handleCancelConfirm}
           confirmLabel="Confirm"
           cancelLabel="Cancel"
           overlayClassName="export-modal-overlay--stacked"
